apps/cv_service/requirements.txt
opencv-python
numpy
torch
pillow
rembg
onnxruntime
matplotlib
segment-anything
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
